feat: moving logout button from profile to menu
