+++ packages/client/package.json
@@ -20,7 +20,7 @@
 		"@auth0/auth0-react": "2.2.4",
 		"@floating-ui/react": "0.26.4",
 		"@tailwindcss/typography": "0.5.10",
-		"@versini/ui-components": "4.19.0",
+		"@versini/ui-components": "4.20.0",
 		"autoprefixer": "10.4.16",
 		"clsx": "2.0.0",
 		"postcss": "8.4.32",
